ai-service/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
mysql-connector-python>=8.2.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.1.0
python-dotenv>=1.0.0
